+++ src/archive/search-result.ts
@@ -2,14 +2,18 @@
 
 /**
  * SearchResult contains information on a scene that may be ordered through the archive API
- * full details can be found in the 
- * @see {https://arlula.com/documentation#search|Arlula documentation}
+ * full details can be found in the Arlula Documentation
+ * 
+ * @see {https://arlula.com/documentation/#archive-search|Archive Search endpoint documentation}
+ * or
+ * @see {https://arlula.com/documentation/#ref-search-result|Archive Search result structure reference}
  */
 export default interface SearchResult {
     supplier: string;
     eula: string;
     id: string;
     sceneID: string;
+    platform: string;
     date: Date;
     center: {
         lat: number;
@@ -33,9 +37,10 @@ export default interface SearchResult {
             additional: number;
         }[];
     };
-    fulfillment: number;
+    fulfillmentTime: number;
     resolution: number;
     thumbnail: string;
     cloud: number;
+    offNadir: number;
     annotations: string[];
 }