+++ src/orders/order.ts
@@ -1,16 +1,15 @@
-import { AxiosInstance } from "axios";
 import Resource, {fromJSON as resourceFromJSON} from "./resource";
 import paths from "../util/paths";
-import { handleError } from "../util/error";
+import { jsonOrError, requestBuilder } from "../util/request";
 
 /**
  * Utility class to construct Orders from JSON data
  * 
  * Note: only to be used internally, not by end users
- * @param {AxiosInstance} client the dialer
+ * @param {requestBuilder} client the dialer
  * @param json the JSON content to attempt to parse into an Order
  */
-export function fromJSON(client: AxiosInstance, json: string|{[key: string]: unknown}): Order|string {
+export function fromJSON(client: requestBuilder, json: string|{[key: string]: unknown}): Order|string {
     if (typeof json === "string") {
         json = JSON.parse(json);
     }
@@ -72,9 +71,11 @@ export function fromJSON(client: AxiosInstance, json: string|{[key: string]: unk
  * @class Order wraps the data that makes up an order
  * 
  * Note: construction of this class is to only be done internally to the library
+ * 
+ * @see {https://arlula.com/documentation/#ref-order|Order structure reference}
  */
 export default class Order {
-    private _client: AxiosInstance;
+    private _client: requestBuilder;
     private _id: string;
     private _createdAt: Date;
     private _updatedAt: Date;
@@ -91,7 +92,7 @@ export default class Order {
      * Create a new order instance
      * @constructor
      * 
-     * @param {AxiosInstance} client    The initiated http transport for the API, created and initialized with credentials by the root Arlula client
+     * @param {requestBuilder} client    The initiated http transport for the API, created and initialized with credentials by the root Arlula client
      * @param {string}        id        The Order ID
      * @param {Date}          created   The timestamp when the order was created (UTC timezone)
      * @param {Date}          updated   The timestamp when the order was last updated (UTC timezone)
@@ -104,7 +105,7 @@ export default class Order {
      * @param {Resource[]}    resources List of resource for this supplier (if available)
      * @param {Date}          [exp]     Expiration date for this orders resources
      */
-    constructor(client: AxiosInstance, id: string, created: Date, updated: Date, supplier: string, imgID: string, scene: string, status: OrderStatus, total: number, type: string, resources: Resource[], exp?: Date) {
+    constructor(client: requestBuilder, id: string, created: Date, updated: Date, supplier: string, imgID: string, scene: string, status: OrderStatus, total: number, type: string, resources: Resource[], exp?: Date) {
         this._client = client;
         this._id = id;
         this._createdAt = created;
@@ -181,24 +182,32 @@ export default class Order {
             return Promise.resolve(this._resources);
         }
 
-        return this._client.get(paths.OrderGet, {params: {id: this._id}})
+        return this._client("GET", paths.OrderGet+"?id="+this._id)
+        .then(jsonOrError)
         .then((resp) => {
-            if (typeof resp.data !== "object") {
+            if (typeof resp !== "object") {
                 return Promise.reject("Order response is not an object");
             }
-            if (!("resources" in resp.data)) {
+            if (!resp) {
+                return Promise.reject("Order response is not an object")
+            }
+
+            // JSON structure not known, use iterable form
+            const r = (resp  as {[key: string]: unknown})
+
+            if (!("resources" in r)) {
                 // order has no resources
                 this.detailed = true;
                 return [];
             }
-            
-            if (!Array.isArray(resp.data.resources)) {
+
+            if (!Array.isArray(r.resources)) {
                 return Promise.reject("Resources is not an array");
             }
 
             const resources: Resource[] = [];
-            for (let i=0; i<resp.data.resources.length; i++) {
-                const res = resourceFromJSON(this._client, resp.data.resources[i]);
+            for (let i=0; i<r.resources.length; i++) {
+                const res = resourceFromJSON(this._client, r.resources[i]);
                 if (!(res instanceof Resource)) {
                     // error in decoding, pass error up the chain
                     return Promise.reject(res);
@@ -209,8 +218,7 @@ export default class Order {
             this.detailed = true;
             this._resources = resources;
             return resources;
-        })
-        .catch(handleError);
+        });
     }
 
 }