+++ src/orders/resource.ts
@@ -1,15 +1,15 @@
-import axios, { AxiosInstance } from "axios";
+import fs, { WriteStream } from "fs";
 import paths from "../util/paths";
-import { handleError } from "../util/error";
+import { requestBuilder, bufferOrError } from "../util/request";
 
 /**
  * Utility class to construct Resources from JSON data
  * 
  * Note: only to be used internally, not by end users
- * @param {AxiosInstance} client the dialer 
+ * @param {requestBuilder} client the dialer 
  * @param json the JSON content to attempt to parse into a Resource
  */
-export function fromJSON(client: AxiosInstance, json: string|{[key: string]: unknown}): Resource|string {
+export function fromJSON(client: requestBuilder, json: string|{[key: string]: unknown}): Resource|string {
     
     if (typeof json === "string") {
         json = JSON.parse(json);
@@ -41,28 +41,46 @@ export function fromJSON(client: AxiosInstance, json: string|{[key: string]: unk
     if (!isResourceType(json.type)) {
         return `Invalid resource type '${json.type}' not recognized`;
     }
+    if (typeof json.size !== "number") {
+        return "Resource size missing";
+    }
+    if (typeof json.format !== "string") {
+        return "Resource format missing";
+    }
+    if (!json.roles || !Array.isArray(json.roles)) {
+        return "Resource roles missing";
+    }
+    if (typeof json.checksum !== "string") {
+        return "Resource checksum missing";
+    }
 
-    return new Resource(client, json.id, new Date(json.createdAt), new Date(json.updatedAt), json.order, json.name, json.type)
+    return new Resource(client, json.id, new Date(json.createdAt), new Date(json.updatedAt), json.order, json.name, json.type, json.size, json.format, json.roles, json.checksum)
 }
 
 /**
  * @class Resource wraps the data that represents an order resource
  * 
  * Note: construction of this class is to only be done internally to the library
+ * 
+ * @see {https://arlula.com/documentation/#ref-resource|Order Resource structure reference}
  */
 export default class Resource {
-    private _client: AxiosInstance;
+    private _client: requestBuilder;
     private _id: string;
     private _createdAt: Date;
     private _updatedAt: Date;
     private _order: string;
     private _name: string;
     private _type: ResourceType;
+    private _size: number;
+    private _format: string;
+    private _roles: string[];
+    private _checksum: string;
     /**
      * create a new resource instance
      * @constructor
      * 
-     * @param {AxiosInstance} client  The initiated http transport for the API, created and initialized with credentials by the root Arlula client
+     * @param {requestBuilder} client  The initiated http transport for the API, created and initialized with credentials by the root Arlula client
      * @param {string}        id      The Resource ID
      * @param {Date}          created The timestamp when the resource was created (UTC timezone)
      * @param {Date}          updated The timestamp when the resource was last updated (UTC timezone)
@@ -70,7 +88,7 @@ export default class Resource {
      * @param {string}        name    An identifiable name/filename for this resource
      * @param {ResourceType}  type    Identifier for the type of resource (imagery, metadata, etc @see ResourceType )
      */
-    constructor(client: AxiosInstance, id: string, created: Date, updated: Date, order: string, name: string, type: ResourceType) {
+    constructor(client: requestBuilder, id: string, created: Date, updated: Date, order: string, name: string, type: ResourceType, size: number, format: string, roles: string[], checksum: string) {
         this._client = client;
         this._id = id;
         this._createdAt = created;
@@ -78,6 +96,10 @@ export default class Resource {
         this._order = order;
         this._name = name;
         this._type = type;
+        this._size = size;
+        this._format = format;
+        this._roles = roles;
+        this._checksum = checksum;
     }
 
      // public getters
@@ -99,23 +121,56 @@ export default class Resource {
     public get name(): string {
         return this._name;
     }
+    /**
+     * @deprecated legacy `resource.type` is deprecated
+     * field will be provided for orders during transition.
+     * 
+     * roles field is the recommended replacement, see
+     * @see {https://arlula.com/documentation/#ref-resource|Order Resource structure reference} 
+     * and the "resource roles" section for details on the new role field and its use.
+     */
     public get type(): ResourceType {
         return this._type;
+    }
+    public get size(): number {
+        return this._size;
+    }
+    public get format(): string {
+        return this._format;
+    }
+    public get roles(): string[] {
+        return this._roles;
+    }
+    public get checksum(): string {
+        return this._checksum;
     }
 
     // actual functions
 
     /**
      * Download the content of a resource (imagery, metadata, etc)
-     * Data is made available as an ArrayBuffer or Buffer depending upon platform.
+     * Data is made available as an ArrayBuffer.
      * 
      * Note: If the order this resource is for has its `expiration` field set and that date has
      * passed, this request will fail as the resource has expired and is no longer hosted in the platform
      * 
-     * @returns {Promise<ArrayBuffer|Buffer>} the content of the resource as a Buffer
+     * @returns {Promise<ArrayBuffer>} the content of the resource as a Buffer
      */
-    download(): Promise<ArrayBuffer|Buffer> {
+    download(): Promise<ArrayBuffer> {
         return downloadHelper(this._client, this._id);
+    }
+
+    /**
+     * Download the content of a resource (imagery, metadata, etc)
+     * Data is piped into the provided fs.WriteStream or one is created at the provided filepath.
+     * 
+     * Note: If the order this resource is for has its `expiration` field set and that date has
+     * passed, this request will fail as the resource has expired and is no longer hosted in the platform
+     * 
+     * @returns {Promise<WriteStream>} file the resource was written to
+     */
+    downloadToFile(ref: string|WriteStream): Promise<WriteStream> {
+        return downloadFileHelper(this._client, this._id, ref);
     }
 }
 
@@ -124,25 +179,61 @@ export default class Resource {
  * 
  * Note: an internal helper not intended to be called directly
  * 
- * @param {AxiosInstance} client The initiated http transport for the API, created and initialized with credentials by the root Arlula client
+ * @param {requestBuilder} client The initiated http transport for the API, created and initialized with credentials by the root Arlula client
  * @param {string}        id     ID of the resource to download
  */
-export function downloadHelper(client: AxiosInstance, id: string): Promise<ArrayBuffer|Buffer> {
-    return client.get(paths.ResourceDownload, {params: {id: id}, responseType: "arraybuffer"})
-    .then((resp) => {
-        if (resp.status === 203) {
-            // fallback for browser where headers would get included in redirect
-            // custom header X-Download-Manual tells the server to send the redirect
-            // url that makes a subsequent request here without the headers configured
-            return axios.get(arrayBufferToString(resp.data), {responseType: "arraybuffer"})
-            .then((r) => {
-                return r.data as ArrayBuffer;
-            })
-            .catch(handleError);
+export function downloadHelper(client: requestBuilder, id: string): Promise<ArrayBuffer> {
+    return client("GET", paths.ResourceDownload+"?id="+id)
+    .then((r) => {
+        const redirect = r.headers.get("location");
+        if (!redirect) {
+            return r;
         }
-        return resp.data as ArrayBuffer;
+        return client("GET", redirect, undefined, undefined, true);
     })
-    .catch(handleError);
+    .then(bufferOrError);
+}
+
+/**
+ * helper function to allow the download call both form a resource and from the orders API client
+ * resources downloaded through this helper are piped into a file rather than an in memory buffer.
+ * 
+ * Note: an internal helper not intended to be called directly
+ * 
+ * @param {requestBuilder} client The initiated http transport for the API, created and initialized with credentials by the root Arlula client
+ * @param {string}        id     ID of the resource to download
+ * @param {string|fs.WriteStream} fileRef File stream or path at which to create the file that the resource will be downloaded to
+ */
+export function downloadFileHelper(client: requestBuilder, id: string, fileRef: string|WriteStream): Promise<WriteStream> {
+    return new Promise((resolve, reject) => {
+        let file: WriteStream;
+        if (typeof fileRef === "string") {
+            file = fs.createWriteStream(fileRef, { 
+                flags: "w+",
+                mode: 0o644,
+            });
+        } else {
+            file = fileRef;
+        }
+
+        client("GET", paths.ResourceDownload+"?id="+id)
+        .then((r) => {
+            const redirect = r.headers.get("location");
+            if (!redirect) {
+                return r;
+            }
+            return client("GET", redirect, undefined, undefined, true);
+        })
+        .then((res) => {
+            res.body?.pipe(file);
+            res.body?.on("error", (e) => {
+                reject(e)
+            });
+            file.on("finish", () => {
+                resolve(file);
+            });
+        });
+    });
 }
 
 /**
@@ -170,7 +261,7 @@ export function downloadHelper(client: AxiosInstance, id: string): Promise<Array
  * License -----------> "license",
  */
 export enum ResourceType {
-    ImageryTiff       = "img_tiff",
+    ImageryTiff       = "img_tif",
     ImageryJP2        = "img_jp2",
     Thumb             = "thumb",
     Overview          = "gdal_tiff_overview",
@@ -188,11 +279,3 @@ export enum ResourceType {
 function isResourceType(token: string): token is ResourceType {
     return Object.values(ResourceType).includes(token as ResourceType);
 }
-
-function arrayBufferToString(buf: ArrayBuffer): string {
-    // std lib typings enforce an argument of 'number[]' and dont handle
-    // the newer numeric array types like Uint8Array 
-    // as a result this converts to unknown then to number array for type purposes
-    return String.fromCharCode.apply(null, new Uint8Array (buf) as unknown as number[]);
-  }
-